Add files via upload
Capstone project
Create a custom Payment Gateway. Third party integrations are not needed. The payment gateway need not be visible on the frontend.
Create a custom Shipping Method. Third party integrations are not needed. The shipping method need not be visible on the frontend
Use the Magento REST API framework to create an order using the created payment and shipping method. 
Invoice and Ship the order using the REST APIs